[GLES2CmdDecoder] Always get size from src texture when copying texture
This CL changes the validating command decoder's CopyTexture() and
CopySubTexture() to always get the size to use from the source texture,
rather than first checking whether there is a GLImage present and
getting the size from that if so. As this change might have unintended
side-effects, we guard it by an enabled-by-default Finch feature. We
also add a flag in about://flags to toggle the feature to make it easy
to determine whether this change is the source of any observed
unexpected behavior.

Our reason for believing that obtaining the size from the GLImage if
present is no longer necessary is that this behavior was introduced for
Android, which no longer uses GLImage at all. Further details in
crbug.com/1412075.

Bug: 1412075
Change-Id: Id52f8639460b7daf1fa89569959fd58350ef7703
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4217451
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100871}