SearchPrefetch: Fail prefetch if no body consumer handle is provided
If the response consumer passed into the `OnReceiveResponse` method
is invalid, we cannot really load the page as there is no response
body, nor would we receive a valid one in the future.
So, a response with an invalid body should not be a servable response.

Bug: 1400881
Change-Id: I9697e4a5518de77492b539e1606de96de9462990
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4211102
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100880}