Remove AutofillPopupBaseView::GetSeparatorColorId.
The separator color has a unique color id for menu separators -
it seems unlikely that we would overwrite it and therefore the
method is not needed.

The two sites where it is used can also be simplified significantly.
There is neither a need for a custom OnThemeChanged method nor
a need for a custom OnPaint method.

Bug: 1411172
Change-Id: I1457d9016aacb39363e897f85cf0a0e7abfb2c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219132
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1100887}