arc: Check battery state before showing ARCVM /data migration screen
When the device does not have enough battery, show a warning and disable
the update button (and also ask the user to connect their device to a
charger if not).

Screenshots:
https://screenshot.googleplex.com/87VrAjHYFDyKAs3 (low storage)
https://screenshot.googleplex.com/9ccCQc44Qn9MWMZ (low battery,charged)
https://screenshot.googleplex.com/B7dReN9bWpiboDJ (low battery,!charged)

Bug: b:258278176
Test: --enable-features=ArcVmDataMigration
Test: Connect/Disconnect the device to/from a charger. Confirm that a
warning to ask charging disappears/appears if it has enough free disk
space but does not have enough battery.
Test: Keep connecting the device to a charger when it has enough free
disk space but does not have enough battery. Confirm that a warning
about the required battery % disappears and the update button is enabled
when the battery % reaches the threshold.

Change-Id: I2209d950cf863103dfb70dba38ed45518255b22f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4049541
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Danila Kuzmin <dkuzmin@google.com>
Cr-Commit-Position: refs/heads/main@{#1100894}