Fix AddressImportRequirement metrics.
On profile import, country-specific completeness requirements are
checked in a function called IsMinimumAddress(). In it, metrics whether
the requirements are fulfilled or violated are emitted.

During multi-step imports, the observed profile is merged with
previously submitted profile fragments to create a complete (w.r.t
IsMinimumAddress()) profile. To do so, the logic reuses the same
IsMinimumAddress() function. As to not emit metrics multiple times
per import (depending on how many merging steps happen),
IsMinimumAddress() has a collect_metrics parameter, that is supposed to
disable the metric collection.

However, in a refactoring (crrev.com/c/3865515) this parameter was
missed and the metrics are currently logged regardless of it. This
causes surprising spikes in the metrics [1] of the multi-step import
rollout.

This CL fixes the bug by checking for the parameter.

[1] https://fleimgruber.users.x20web.corp.google.com/uma_metrics/AutofillEnableMultiStepImports-2023-01-28-14days-STABLE-3/index.html#WIN-HISTOGRAMS

Bug: 1313367
Change-Id: Ic8b06cc0e91afaeefb9db10ea661d1a4a39b61a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209628
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1100898}