[ios] Generalize update_ios_bundle_data.py and add PRESUBMIT checks
With this patch, machinery has been added to populate file lists
based on a list of globs, generalizing the approach taken from
update_ios_bundle_data.py. It is also possible to check that these
file lists are kept up to date via PRESUBMIT checks. A support
utility has also been added to simplify the creation of new presubmit
checks. Details on the conventions used in the glob lists are
outlined in update_bundle_filelist.py.

A number of large bundle_data's have been updated in this patch
but there should be no behavioral change.

Regarding rationale: it can be cumbersome to maintain large sources
lists in bundle_data, and while there was a script in //net to
simplify updating some net-specific lists, it wasn't available
elsewhere, nor was there a PRESUBMIT check to remind you that an
update is needed (or how to do it). This patch makes this approach generally available and provides guidance via PRESUBMIT.

credit: thanks to sdefresne@ for suggesting the read_file, PRESUBMIT,
and bundle_data_from_filelist.gni approaches and to dtapuska@ for
suggesting the presubmit_support.py to simplify the various
PRESUBMIT.py files.

Bug: 698042
Change-Id: Ia5e79432ab392ad3530e92cb76de54b012dfe5d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209248
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100913}