[Kiosk] Remove SetIgnoreKioskAppDataLoadFailuresForTesting
The method KioskAppData::SetIgnoreKioskAppDataLoadFailuresForTesting was
a workaround to let kiosk apps be installed in tests while we fail to
fetch their details (outside of tests this is done via a request to
Chrome Web Store.

This change adds a minimal setup to affected tests to handle this
request, after which we don't need a special test-only method which
disables all the checks.

Bug: b:246912087
Change-Id: I86a2456d565621bbbfca4adbb9adeb90e1fd10c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4208934
Commit-Queue: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Ivan Šandrk <isandrk@chromium.org>
Reviewed-by: Ben Franz <bfranz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100971}