Fix test-only UAF in PrefServiceSyncableMergeTest.ManagedListPreferences
This test passed a pointer to a tightly-scoped local variable to the
InitWithSyncDataTakeOutput() helper - which keeps the pointer around.
Once the local variable went out of scope, the pointer became dangling,
causing potential use-after-free issues.

This CL quick-fixes the problem by extending the scope (and thus
lifetime) of the local variable. That brings this particular test in
line with all the other tests in this file.
Longer-term, a better solution would be to avoid the dangerous pattern
of passing out a raw pointer.

Bug: 1404937
Change-Id: I0ed708cf2f623c50495f34fcb5c95b57378b0f29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4221232
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101007}