[omnibox][grouping] Verify SideType/RenderType of Groups in Sections
Verifies at construction time whether the Groups in a given Section
have the same SideType (Primary/Secondary) as well as the GroupIds
permitted in a given Group have the same RenderType
(Horizontal/Vertical).

In order to verify this at construction time a Section now receives a
list of Groups in its constructor. And in order to store a list of
Groups instead of unique_ptrs to Groups this change folds DefaultGroup
into the base Group. This makes the configuration of the default group
explicit which is probably less error prone and makes it possible to
store Section configurations in a text format for improved readability
and referencing purposes.

Bug: 4123156
Change-Id: I3c941d9c184476b0e65a5e8b307b18d8ca0f5d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4129781
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101034}