[ChromeVox] Panel shows on primary display properly
Previously AccessibilityController assumed that the panel was only
shown on the primary display, but it was actually being added to
the display for new windows. Updates the panel to be added to only
the primary display.

This way the panel is always shown no matter which display is active.

AX-Relnotes: Fixes bug where ChromeVox panel didn't show up on second display
Bug: b:260181195
Test: New automated test which previously would have failed; manual
Change-Id: I7da1a89a4c8ed1682c820f2e39639c564272d1c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220131
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101052}