Add PostPresentation breakdown to EventLatency
It is possible for swap-end to happen after presentation. In this case,
SwapStartToSwapEnd (or LatchToSwapEnd if swap breakdowns are available)
as a nested breakdown of the top-level EventLatency event, won't fit
properly in the EventLatency event causing Perfetto show the entire
slice incorrectly. To fix this, we divide EventLatency into two
breakdowns: EventLatency.ToPresentation and
EventLatency.PostPresentation. The to-presentation breakdown will
truncate its nested slices to presentation. If there is any slice that
gets truncated, it will be shown under the post-presentation breakdown.
See full details of the design in
https://docs.google.com/document/d/1ZUr_TWS2bMF0uWq5PqTjMlz1RAxBpZOIyRoSzOeTGBA/edit?usp=sharing&resourcekey=0-1zgv2tE4Oj9KuagOXWsMnw

Note that this further breakdown is only performed on EventLatency
events that are associated with a successfully presented frame. If the
event is not associated with a frame, we won't have this issue, so we
don't add this breakdown at all.

Bug: b/259434854
Change-Id: I4f490fa4f17e3cb2db004721355c37c0c6bf1dee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4197826
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101054}