Add a few explicit conversions to size_t in device/fido.
base::span will shortly require that length is unsigned.

This CL fixes two minor issues in device/fido related to this
upcoming change.

1. SHA256_DIGEST_LENGTH, as defined, is unfortunately untyped
so a static cast is added in this CL to make it a size_t
correctly so it can be passed into make::span.
2. In virtual ctap2 device, GetUnsigned returns a uint64_t which
might not fit into a size_t on 32-bit platforms. Thus a
checked_cast is added to handle this.
3. The manual pointer arithmetic to obtain the response key is
replaced by a call to subspan.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: Id77471b96214061994dcc4c13c96171dce41f722
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220318
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101081}