Ensure spans are created with unsigned length in cast test simulator.
An upcoming change will enforce that base::span sizes are unsigned
values and fit into size_t type.

AppendYuvToFile was previously taking signed stride and rows and
passing those into base::span. This will no longer be allowed
in an upcoming CL.

This CL adds a series of checked_casts to make sure that stride
and rows does not overflow when multiplied, and also that both
stride and rows are non-negative values, before passing to the
base::make_span function.

This means that the code will continue to compile once the
base::span checks have landed in a future CL.

No functional changes are intended with this CL.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: I609b56280726e9ba4dd66b288be370ef1de78633
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4217484
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101088}