[Cast Remoting] Fix AV Sync When Audio Disabled
When audio is disabled, the transition from remoting back to mirroring
results in increased playback delay, as the mirroring session will wait
for audio the audio_buffering_state to change to BUFFERING_HAVE_ENOUGH,
which takes a few seconds.

This CL disables audio in the mirroring/remoting session in such cases
when audio is unsupported.

Bug: 1403250
Change-Id: Ifa66a5950c5c83847851fc568ed76dc71901d956
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210948
Commit-Queue: Ryan Keane <rwkeane@google.com>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101138}