Use the host's scale factor if the client's scale factor is close.
A client is expected to send the same value as exo uses. However this
is not guaranteed because walyand doesn't support float as data type,
and existing impl uses logical size + pixel size to compute the
scale factor, which can have an error. (and this becomes bigger when
display zoom is applied).

We'll work on a new extension protocol to fix this (1412420), and
meanwhile, this CL will mitigate the issue by using the same scale
factor as host when the value is close.

This also updates the logic to compute the scale factor + zoom factor
so that the logical size x scale factor will produce near integer
size.

Bug: 1411793, 1412420
Test: DisplayResolutionTest.DisplayZoom

Change-Id: I73ffb1712bb51a4fc5a27bf0c5609e87a781dc98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210142
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101154}