[ntp][side-panel] Fix uploading an image on NTP when CWS theme is set
Problem: Themes aren't being disabled when an uploaded image set.

Solution TLDR: We need to disable the extension when we upload a background image. We currently do it for Classic Chrome and 1P themes using two different methods. We can use one of the two existing ones or create a third for this case. This CL proposes using the existing method we use for Classic Chrome - disabling the extension and removing the color that is set.

Solution: Classic Chrome disables extensions by calling "UseDefaultTheme" in theme_service.cc[1]. This removes the current color that is set as well. Chrome Colors and 1P themes disable the extension and build a theme from a color using "BuildAutogeneratedThemeFromColor" in theme_service.cc[1]. This adds a new color when the extension is disabled. Both of these result in calls to DisableExtension() in theme_service.cc[1].

Uploading an image has neither of these calls. We either need to mimic the calls from one of the above two options or implement a new function to disable an extension directly. This CL proposes removing the color as well as the extension to let the user know their extension has been removed completely.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/themes/theme_service.cc

*Video of observed changes is in bug comments*

Change-Id: I16c182b8b194aab225b12aa094bb74f7cfdd20b8
Bug: 1411817
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219633
Commit-Queue: Paul Adedeji <pauladedeji@google.com>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101156}