Shared Storage: Change default value of IsSharedStorageAllowed
We want to default `ContentBrowserClient::IsSharedStorageAllowed()` to
false, but need to be able to override this value for tests.

Overriding is relatively straightforward to do for both browser tests
and wpt_internal/ web tests (see PS 4), but the extra work needed for
external/wpt/ tests is harder to justifying doing for the few web
tests that would currently need it.

So instead, we improve the status quo slightly by defaulting to false
in `ContentBrowserClient` and making an override to true in
`ShellContentBrowserClient`, for both the aforementioned method and
its sister method `IsSharedStorageSelectURLAllowed()`.

Bug: 1325103,1218540
Change-Id: I8fc95002da96a6255f09228def92434daf1392c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220158
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101180}