[Eche] Create mojo interface for orientation changes
This change introduces a mojo interface that allows the web UI to signal
to the bubble to switch between horizontal and vertical orientations.
Follow-up CLs will handle the usage of the signal from the web UI side
and the orientation change on the C++ views side.

Test: verified that Phone Hub and App Streaming work properly.
Bug: b/258306301
Change-Id: I56c2dca7305068c3b95649c0937257da9ef780ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4202054
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Abbas Nayebi <nayebi@google.com>
Commit-Queue: Crisrael Lucero <crisrael@google.com>
Reviewed-by: Jon Mann <jonmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101207}