Set out-of-bounds printable areas to be the paper size
Some printer drivers have Papers with printable areas out of the bounds
of the paper size. This caused Mojo validation errors, since it
originally marked out-of-bounds printable areas as invalid. Instead, set
the printable area to be the paper size and mark it as valid.

Also set the printable area to be the paper size in Windows if the
printable area is empty.

Fixed: 1412305
Change-Id: I608a5661a0729af498f7e67308fa6410ab6ced84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219867
Reviewed-by: Alan Screen <awscreen@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101298}