Don't create queue with corrupt file paths
Sync changes from https://crrev.com/c/4211444.
This change returns an error status for the StorageQueue constructor if
a generation id is not found in any of the file paths in the directory
assigned to the StorageQueue. Previously we attempted to recover this
data using metadata files, but if the generation id is missing from every single data file path, it's safe to assume that the entire directory is corrupt and should be deleted.

Change-Id: I2ee61e75b7d768d379d63a6c54da0d12b9391a9e
Bug: b:267429525, b:267497955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219002
Commit-Queue: Josh Hilke <jrhilke@google.com>
Reviewed-by: Hong Xu <xuhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1101322}