Check for both false values and for unset values
The previous code was incorrect and would fail if the value
was false, since if you check the value of an optional bool
that is false, it returns true if it is not null.

I want this to pass if the value is false.

This triggered the crash described in b/267714781.

BUG=b/267714781

Change-Id: I837d092c236c2435bda82e5caefe217f931bdd59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4223596
Reviewed-by: Darren Shen <shend@chromium.org>
Auto-Submit: Keith Lee <keithlee@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101417}