[conflict dialog] implement CopyOrMoveIOTask pause and resume
The default behavior of files app today is to 'keepboth'. When the
feature flag is off or the user chooses to 'keepboth', then source
entry 'test' is named 'test (2)` at the destination (keepboth) and
no change in behavior. Existing tests cover this behavior:

 KeyboardOperations/FilesAppBrowserTest.Test/keyboardCopyDownloads
 KeyboardOperations/FilesAppBrowserTest.Test/keyboardCopyDrive

and run with conflict dialog feature flag on/off crrev.com/1093196
for reference.

The fact that the suggested destination name is 'test (2)' and not
'test' indicates that 'test' exists at the destination viz., there
is a file name conflict. And this is a-priori known before we copy
or move an item in the operation: GenerateDestinationURL generates
the destination name for the current source item (at |indx|), then
calls CopyOrMoveFile (at |indx|) with that destination name.

CopyOrMoveFile can examine destination name to figure out if there
is a file name conflict and change the code flow if so: enter task
state::PAUSED, and send Pause params to the UI to resolve. Comment
all the "figure out" steps in detail.

The UI asks the user to resolve the file name conflict and Resumes
the copy_or_move with the Resume params, which contains 'keepboth'
or 'replace'. For a 'replace', delete the destination entry first,
and call ContinueCopyOrMoveFile. For a 'keepboth' we can just call
ContinueCopyOrMoveFile directly.

The Resume params also contain an "apply_to_all" field that can be
used to resolve all future file name conflicts in the copy_or_move
operation in the same way (no need to ask the UI again).

Bug: b:255264604
Change-Id: Ic92ae06f90a6e112e68ab897012a968976bfe1b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4216435
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101421}